画像を確認する部分を別コンポーネントに切り出し
